requirement.txt
flask
flask-cors
python-dotenv
openai
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken
youtube-transcript-api
